+++ my-app/src/app/form/observer/page.tsx
@@ -1,0 +1,5 @@
+import FormPage from "../observerForm";
+
+export default function ObserverFormPage() {
+  return <FormPage />;
+}
