+++ my-app/src/app/form/page.tsx
@@ -1,4 +1,4 @@
-import FormPage from "./form";
+import FormPage from "./dspForm";
 
 export default function Page() {
   return <FormPage />;
